+++ compose.yml
@@ -54,7 +54,6 @@ services:
       - node_modules_ghost_i18n:/home/ghost/ghost/i18n/node_modules:delegated
       - node_modules_ghost_in-memory-repository:/home/ghost/ghost/in-memory-repository/node_modules:delegated
       - node_modules_ghost_job-manager:/home/ghost/ghost/job-manager/node_modules:delegated
-      - node_modules_ghost_link-redirects:/home/ghost/ghost/link-redirects/node_modules:delegated
       - node_modules_ghost_link-replacer:/home/ghost/ghost/link-replacer/node_modules:delegated
       - node_modules_ghost_magic-link:/home/ghost/ghost/magic-link/node_modules:delegated
       - node_modules_ghost_mailgun-client:/home/ghost/ghost/mailgun-client/node_modules:delegated
@@ -195,7 +194,6 @@ volumes:
   node_modules_ghost_i18n: {}
   node_modules_ghost_in-memory-repository: {}
   node_modules_ghost_job-manager: {}
-  node_modules_ghost_link-redirects: {}
   node_modules_ghost_link-replacer: {}
   node_modules_ghost_magic-link: {}
   node_modules_ghost_mailgun-client: {}
