+++ ghost/core/core/server/services/link-redirection/RedirectEvent.js
