+++ ghost/core/package.json
@@ -103,7 +103,6 @@
     "@tryghost/kg-lexical-html-renderer": "1.3.5",
     "@tryghost/kg-mobiledoc-html-renderer": "7.1.1",
     "@tryghost/limit-service": "1.2.14",
-    "@tryghost/link-redirects": "0.0.0",
     "@tryghost/link-replacer": "0.0.0",
     "@tryghost/logging": "2.4.21",
     "@tryghost/magic-link": "0.0.0",
