--- ghost/link-redirects/lib/link-redirects.js
@@ -1,5 +1,0 @@
-module.exports = {
-    LinkRedirect: require('./LinkRedirect'),
-    RedirectEvent: require('./RedirectEvent'),
-    LinkRedirectsService: require('./LinkRedirectsService')
-};
